fix ModelUtils.getAdditionalProperties()
This commit fixes the ModelUtils.getAdditionalProperties() function
so that if a schema contains:
    additionalProperties: true
the function would return an empty schema, which implies that
additionalProperties can be ANY type.
Previously, the function would return an empty OBJECT schema
which implies that additionalProperties can be an object
with any keys and values, a subtle but important distinction.